Prevent Linux OS runtimes on WSL

+++ .shell/os_linux.sh
@@ -2,7 +2,7 @@
 # Linux specifics
 #
 
-if [[ "$OSTYPE" =~ ^linux ]]; then
+if [[ "$OSTYPE" =~ ^linux ]] && [[ -f /proc/version ]] && grep -v --quiet Microsoft /proc/version ; then
 
   # Link VSCode config
   ln -sf "$HOME/.vscode/settings.json" "$HOME/.config/Code/User/settings.json"
